requirements.txt
# Requires Python >= 3.9
requests==2.32.5
pyyaml==6.0.3
InquirerPy==0.3.4
